+++ AudioOutput.h
@@ -9,7 +9,7 @@
 class AudioOutput : public AudioModule {
 public:
     AudioOutput();
-    void process(Uint8* stream, int length) override;
+    void process(Uint16* stream, int length) override;
     void render() override;
     std::vector<ed::PinId> getPins() const override;
     ed::PinKind getPinKind(ed::PinId pin) const override;
